fix test_download_multiple
This test was trying to verify that if two download locations are provided,
both are recorded in the workspace_rules_log. It also tested that the
build completed successfully; however, the latter only happened by accident,
as the testing server would blindly return 200 and only then try to read
the file---and then crash, as file1.txt does not exist; so the file returned
had to relation at all to the file served at the second (valid, but never
accessed) URL. This was uncovered when adding a checksum in preparation of
the upcoming #8607 flag flip.

As the test is only about logging, put the use of download to the intended
form that all provided URLs are valid and return a file with the correct
hash.

Change-Id: I208ebffc5a724df7a75b2bced6fafbd8f50417eb
PiperOrigin-RevId: 259486544