Stream QueryResult contents in query --output=proto
Emits targets as they arrive instead of accumulating them in a QueryResult.
Based on some simple benchmarks this speeds up end-to-end query time by up to
7%. This should also help with genquery, especially those with large results.

Internally this is implemented by using knowledge of the proto serialization
format and the simplicity of QueryResult to skip the intermediate
QueryResult.Builder and QueryResult. On one hand this can be a somewhat
questionable blurring of abstraction layers, but on the other QueryResult's
simplicity and all of our test coverage for it makes it a safe, fair tradeoff.

PiperOrigin-RevId: 259603992