register_{toolchains, execution_platforms}: honor renaming
When handling the registration of toolchains or execution platforms,
take the applicable renaming into account. This, in particular, includes
the renaming from the name of the main workspace to the canonical '@'.

Fixes #7773.

This is a roll forward of be25bbf38d9337432c1e79e3f027ad5d0122bb40
now that #8937 is fixed. Also include a regression test for #8937
(this test couldn't be added in the fix for #8937, as the example
depends on #7773 being fixed first).

Change-Id: I24d91275531218db408f612a54bde928c1a0b3db
PiperOrigin-RevId: 259952821