Factor out filtering functions from universe in rdeps(u, x, 1).
For example, the query `rdeps(kind('binary', ...), //foo:bar, 1)` will be
first rewritten as `kind('binary', rdeps(..., //foo:bar, 1))`.

This rewriting is not much of an optimization on its own, but it may make the
query eligible for the RdepsToAllRdepsQueryExpressionMapper, which can rewrite
the example intermediate query as `kind('binary', allrdeps(//foo:bar, 1))`.

RELNOTES:
None.
PiperOrigin-RevId: 260773356