Remove ArtifactFileMetadata and replace it with FileArtifactValue.
A harmful side effect is the loss of the "digest must be present" assertion on RegularFileArtifactValue. I'm not a fan of this, however, in practice:

- There are other kinds of FileArtifactValues and it's not clear whether they are supposed to have digests or not
- The important invariant, that ActionExecutionValue supplies a digest in all sane cases, is not upheld at HEAD because ArtifactFileMetadata is not guaranteed to contain one.

The game plan is to re-instate that check when ActionExecutionValue is constructed (somehow) after getting rid of the PLACEHOLDER silliness, if possible.

In exchange, however, we get a saner treatment of output directories: before this change, we created ArtifactFileMetadata instances for them even though they were not, well, files.

RELNOTES: None.
PiperOrigin-RevId: 260909718