Refactor SimpleBlobStore#put
Split it into SimpleBlobStore#uploadFile and SimpleBlobStore#uploadBlob.
The return type changes to ListenableFuture<Void> but all
implementations are still blocking for now. This will be changed in a
follow up CL.

This is a refactoring towards the larger goal of removing SimpleBlobStoreActionCache
and making GrpcRemoteCache a SimpleBlobStore implementation. The idea
is for AbstractRemoteActionCache#upload(ActionResult, ...) to no longer
be abstract but instead to call out to abstract uploadFile and
uploadBlob methods (just like downloadBlob).