Remove {OutputStore,ActionExecutionValue}.additionalOutputData.
Instead of having another map, when additional data is needed, replace artifactData with it.

There was one place where we exploited the difference between the two: for cases where we get a digest from remote execution but one is not available for the file system, FilesystemValueChecker used only one map.

I fixed this by adding a method to FileArtifactValue that tells whether given two instances, Blaze should assume that there was no change and replicated the existing behavior as faithfully as possible.

RELNOTES: None.
PiperOrigin-RevId: 261086332