Get ActionAnalysisMetadata in a simpler, more efficient manner.
Instead of checking whether it's an action or a template and calling the corresponding method, when it can be either, simply bypass checks and get it directly from the list.

The actions list itself used to not be exposed, so isActionTemplate was necessary to determine which method to call (getAction or getActionTemplate).

Also prevent boxing of ints in the Preconditions calls.

PiperOrigin-RevId: 261174117