Don't special-case caffeine in JavaBuilder's annotation processor classloader
the class prefixes here are things that error prone plugins need to be
able to load from JavaBuilder's classpath: the Error Prone API, and other
types referenced directly in the API (e.g. guava).

As long as caffeine is being used as an implementation detail of
JavaBuilder and does not appear in APIs used by plugins, this is
unnecessary.

PiperOrigin-RevId: 261366846