Clear marker file on repository override
Bazel internally uses a marker file to check if the definition
of an external repository has changed. If an external repository
is overridden by a command-line flag, we cannot just keep the
old marker file, as on the next call without the override, the
effect of the override has to be undone. Therefore, clear the
marker file, if an override is active for the given repository.

Fixes #8543.

Change-Id: I775791c45da9321619b6dc2c7477c645852a6f99
PiperOrigin-RevId: 262086344