Store the total number of runs per test
Instead of recomputing the runs per test everywhere it's needed, compute
it once and store it in the TestTargetExecutionSettings.

This is purely a cleanup change and should have no impact on semantics.

This should not affect memory consumption, because object size is a
multiple of 8, and there were previously 11 4-byte fields (assuming
compressed oops) in this class.

PiperOrigin-RevId: 262096308