Log if attempting to rewind a lost source file input
There's nothing to rewind if the lost input is a source file. If this
occurs then there's a bug in action execution machinery.

This logging replaces what would be an uninformative class cast
exception, and is like the logging done for lost input *owners* if they
(unexpectedly) are source files.

RELNOTES: None.
PiperOrigin-RevId: 262170550