Add a 'local_defines' attribute to cc_* rules
Currently, the only way to specify non-transitive defines is through the 'copts' attribute, which is not platform independent. the cc_*.local_defines has the task to pass define values to the compile command line to the given cc_* rule, but not to its dependents.

Fixes #7939

RELNOTES: cc_* rules support non-transitive defines through a 'local_defines' attribute.
PiperOrigin-RevId: 262315252