Inject ResourceManager everywhere
This is in preparation for async local execution, for which we require a
thread pool. My plan is to have the thread pool be part of
ResourceManager, and injecting it avoids having a long-lived thread pool
in Bazel. Even if we didn't inject the RM, we'd still have to inject the
thread pool, so might as well do it now.

PiperOrigin-RevId: 262519813