Rewrite AbstractSandboxSpawnRunner to use Subprocess
It was previously using Command. This change is in preparation for
merging the local sandboxed and non-sandboxed execution paths, in order
to simplify supporting async execution for both.

I copied over the equivalent code (with minor changes) from
LocalSpawnRunner with the intent of unifying the LocalSpawnRunner and
the AbstractSandboxSpawnRunner. It may be possible to extract this into
a common class rather than having two copies. TBD.

PiperOrigin-RevId: 262545190