Add support for unresolved symlinks as outputs of actions.
Implemented by adding a bit to Artifact that tells if it is an unresolved symlink. In the latter case, the digest of the artifact is computed from the return value of readlink() and a little tweaking to avoid hash collisions instead of resolving the symlink and using the checksum of its eventual target.

This bit is also added to ActionInput because execution strategies also need to know it (for example, when a sandbox strategy encounters a symlink, if the symlink is unresolved, it creates it within the sandbox verbatim, otherwise, it chases it and creates a file with the same contents as the target of the symlink)

Limitations:

- Does not work with RBE (this is doable with some coding, but I'm not familiar with the code base)
- Does not work with sandboxfs or the Windows sandbox

Other than these issues, this seems to work reasonably well. In particular, action caching works as expected (and is tested)

Progress towards #9075.

RELNOTES: None.
PiperOrigin-RevId: 262950280