Bazel server won't crash for bad flag value
Two of the first things the Bazel server does in
BlazeCommandDispatcher.execExclusively is parsing
command line options and initializing modules.
(Modules are Bazel-internal components.)

Parsing the flags consists of two phases:

1. Tokenizing the flags and storing the raw values
   as strings. This step is done eagerly.

2. Parsing values out of the strings and assigning
   them to typed class members. This step is done
   lazily when something asks the OptionsParser to
   return a specific Options instance.

Bazel catches bad flag values in step 1 already
and stores an error code, but it carries on with
initializing all modules. Bazel tries to
initialize all modules, even if some of them fail
to do so, and only then will Bazel observe the
error code.

This is the problem though: if a module that's
being initialized tries to read an option that got
an invalid value (like SandboxModule trying to
read SandboxOptions.sandboxDebug, but
"--sandbox_debug" got "/foo" as a value), then the
OptionsParser will fail to parse the right value
and fail to assign it to the sandboxDebug field,
and Bazel crashes.

This PR makes Bazel silently ignore the failure to
set the value. This is fine: Bazel still reports
the bad flag, but it also leaves the field at its
default value.

Fixes https://github.com/bazelbuild/bazel/issues/7847

Change-Id: I5b20332af57eaad8106018412ab06d4507468736

Closes #9145.

Change-Id: I4cb54c2ad0b94949efc67f6a852d6baad90c1309
PiperOrigin-RevId: 263725800