Ext.repos: improve error message for missing path
If Bazel fails to fetch a local repository because
the user-defined "path" is wrong, the error now
references that user-defined path, not the
Bazel-computed absolute path.

Example:

    local_repository(
        name = "foo",
        path = "./bar",
    )

For Bazel this means "bar" under the workspace
directory. If that does not exist, then Bazel
reports an error.

Until now, this error referenced "bar" by the
computed absolute path, which is sometimes
confusing.

Now Bazel reports the path as it was defined in
the WORKSPACE file, or as defined by the ANDROID_*
envvars.

Affected rules:
- local_repository
- new_local_repository
- android_sdk_repository
- android_ndk_repository

Fixes https://github.com/bazelbuild/bazel/issues/9139

Change-Id: I6df8708c65ce593d8b60b806bfe6c6522da1a51f

Closes #9161.

Change-Id: If9e081bb826f7cc1a163f199eb5e08f0807d7b72
PiperOrigin-RevId: 263793064