Stop rewriting invalid resource qualifiers when building with AAPT2
This makes good on the "This will be an error in the future." warning.

The rewriting for AAPT1 (in FullyQualifiedName.java) is left as-is for
legacy apps, but there's no reason to carry that into the AAPT2 world.

Both AAPT and AAPT2 correctly reject these:
  $ aapt package -F foo.apk -M AndroidManifest.xml -S res/
  invalid resource directory name: res/ values-es-419
  $ aapt2 compile -o out res/values-es-419/strings.xml
  res/values-es-419/strings.xml: error: invalid configuration 'es-419'.
so there was never a good reason for this hack.

BCP47 for resource qualifiers is documented at:
https://developer.android.com/guide/topics/resources/providing-resources.html#table2

The code from ResourceCompiler.java would have omitted translations if both
values-es-419/ and values-b+es+419/ exist, which basically turns what would
have been a build-time error into something that manual QA would need to catch.

RELNOTES: N/A
PiperOrigin-RevId: 263880196