Initial checkin of exec-server command
This command is still a work in progress, is not documented, and is not
officially supported - don't start relying on it unless you have explicit
permission or you really know what you're doing.

There are plenty of rough edges, but getting the bare minimum out to allow me
to start parallelizing remaining work and sanding rough edges.

PiperOrigin-RevId: 264476191