Create proto file containing server metadata on server startup
This is a combination of all of the other server files (except commandline).
Meant to be used with exec-server to determine when the server is online and
ready. It might also be useful to remove the individual info files
(command_port, request_cookie, response_cookie) in favor of this one.

RELNOTES: None
PiperOrigin-RevId: 264650984