Move some of the Build Encyclopedia documentation to a separate page
Build Encyclopedia is a page that is hard to maintain and update. It is partially generated and mixes Google-specific content with open-source content.

I suggest we remove the Build Encyclopedia page from the website. Instead, we can maintain a simple static Markdown file. This may allow us to write more user-friendly content and blur the line between native rules and Starlark rules.

Future changes:
 - Update the menu (after a release)
 - Add the links to "Concepts and terminology" and "Functions" somewhere else
 - Iterate on the "Rule" page's content

RELNOTES: None.
PiperOrigin-RevId: 264659282