Fix crash when test_suite includes test in broken package, with -k
In --keep_going mode, Bazel would crash if a test_suite rule referred to
a test in another package when that other package was unloadable.

Now it emits the package errors as expected and does not crash.

Drive-by fix to adjacent integration test to cover both --keep_going and
--nokeep_going modes.

Fixes https://github.com/bazelbuild/bazel/issues/6586

PiperOrigin-RevId: 264848847