repository mapping lookup: convert to canonical name first
By the way we propagate the information in which repository we are in,
we always have a repository name that is valid in the main repository;
however, it is not necessarily a canonical name. Our map of the repository
mappings, however, is keyed by canonical names. So, add the missing indirection
and first look up the canonical name for the repository name we're given
by looking it up in the mapping of the main repository (for which we know
its canonical name) and only then use that name as key.

Fixes a bug related to #7130.
Prerequisite to flip --incompatible_remap_main_repo.
Therefore, consider for Bazel 0.29.0 (#8572).

Change-Id: Icd426a18aaa406b614f4948a8122177463a72959
PiperOrigin-RevId: 265012106