Use mkdtemp to create cas_path
Something is still interfering with the previous approach, causing the
CAS path to already exist.  I don't know what it could be, but let's
sidestep the issue entirely by using mkdtemp, which will create the
directory itself and pick something such that it doesn't already exist.
This does carry the risk of a slightly longer path, but I suspect it
will still be short enough to work okay.