Lift non-core methods out of OutputFormatter
About 80% of OutputFormatter (by LOC) was utility methods that weren't part of
the actual interface. Move these to proper classes and utility classes to make
it easier to understand exactly what's expected of OutputFormatters by skimming
the class, and should make it easier to test the individual utilities.

RELNOTES: None
PiperOrigin-RevId: 265580839