exec-server fails if a server already exists
exec-server is intended to be used by users who want to maintain full control
of a servers lifecycle. By refusing to start when a server is running we're
making sure that users are keeping up their end of the bargain, as well as
making sure that we're we're not clobbering another server on the user's
behalf.

RELNOTES: None
PiperOrigin-RevId: 265587273