Have exec-server require that install and output base are set
The intention is that users are extra hands-on with server lifecycle and
environment management. This helps enforce that.

I would have liked to have done this in a cleaner/more principled way (ie by
not checking the command name inline), but the way we mutate startup options
and the work I'd need to do to work around it was becoming more than I'm
currently willing to take on.

PiperOrigin-RevId: 265815618