preamble to Starlark rule documentation: mention label
Stardoc, as opposed to skydoc, ignores module doc-strings. Therefore,
add at least some basic information to the preamble that is added to
the generated markdown document; as that preamble is shared, we have
to keep the text generic, hence, for the time being, only mention the
label the files can be loaded from.

Fixes #9225.

Change-Id: I6b19cd8b5f68e9726c0a113a261e2064302152c1
PiperOrigin-RevId: 266356428