Have remote_outputs=toplevel download runfiles. Fixes #8899
This change makes it so that when building a binary remotely and
having --experimental_remote_download_toplevel set the runfiles of the
binary will also be downloaded. This applies also to building a test
target. However, when executing a test remotely with bazel test neither
the test binary nor its runfiles are downloaded. Instead Bazel fetches
the test result: test.xml and test.log

This change also enables bazel run to be used with
--experimental_remote_download_toplevel.