Call FindMissingBlobs before BES artifact upload. Fixes #8673, #7630
Bazel unconditionally uploads files referenced by BEP events to the
remote cache. This change makes it so that Bazel calls
FindMissingBlobs() before uploading a file.

This should drastically improve BES performance for when files already
exist remotely (i.e. when using remote execution).

A side effect of this change is that it also moves calls to potentially
blocking Path#getDigest() on the thread pool.

Co-authored-by: Alexander Lyon <alexander.lyon@codethink.co.uk>