bazel syntax: move SkylarkCallbackFunction to packages.StarlarkCallbackHelper
By moving this type, we can replace StarlarkContext with BazelStarlarkContext,
which removes one more reference to StarlarkContext from lib.syntax.

The new class name reflects that its instances are not functions.
("Helper" is a technical term meaning "bad abstraction".)

This CL also tightens the parameter types of
rule.implementation, aspect.implementation, and rule.outputs
so that they accept only Starlark ('def') functions, not
callable built-ins such as rule classes, providers, etc.
Strictly speaking this is an incompatible change, but it is
almost inconceivable that anyone outside Blaze itself is
exploiting it.

PiperOrigin-RevId: 267241938