bazel syntax: remove StarlarkContext parameter from Starlark intrinsics
Before, the StarlarkQueryable.containsKey and StarlarkIndexable.getIndex
methods had StarlarkContext parameters, for the sole benefit of
ResolvedToolchainContext, which needs to apply the contextual repo
mapping to any string-form labels it encounters in Starlark "str in rtc"
or rtc[str] operations.

This change causes the ResolvedToolchainContext to require a repo mapping
at construction so that we can simplify the Starlark interpreter API,
and eventually get rid of StarlarkContext.

PiperOrigin-RevId: 267292895