bazel packages: move lib.syntax.Type here
Type is a concept of the build language, not of core Starlark.
(Witness all the references to Label. That said, there appear to be
three concepts---BuildType, Type, StarlarkType---where only two
are necessary: the types of rule attributes, and a library of
Starlark-to-Java conversion combinators.)

This change is mostly mechanical, except:
- MethodLibrary: two uses of Type.XYZ.convert replaced by ad-hoc code.
- SkylarkDict: getDictFromArgs substantially rewritten to avoid Type.
  Its error messages were improved.
  I added comments re: harmful type parameterization.

This CL depends on unknown commit, which first breaks
Copybara's dependency on syntax.Type.

PiperOrigin-RevId: 267694761