Move the field name back into FieldInitializer
There isn't a good reason to decouple a field name from its value, unless if
one were to:
* swap out a value for an existing name
* write the same value multiple times with different names
The format was done implicitly as a workaround (b/140643407), and the latter,
if done as an optimization, belongs in aapt2.

On the other hand, we need to attach additional metadata for an R field to
track its provenance, and having name/value be separate makes it unclear where
to put the additional metadata.

This undoes unknown commit.

PiperOrigin-RevId: 267884392