Sanitize strategy creation for better readability.
Make the createSpawnStrategy a static function so that understanding what's
going on in each test only requires reading that test, not the whole class.

Yes, this results in slightly longer tests, but the increased readability
is worthwhile. In particular, this enables another follow-up cleanup that
will encapsulate the behavior of each mock strategy further.

RELNOTES: None.
PiperOrigin-RevId: 268447544