Fix crash with async tree deletions if the executor has not yet initialized.
By the time afterCommand runs, we don't know if executorInit has run so we
cannot make any assertions about the state modified by the latter.  The code
contained one such assertion that relied on the asynchronous tree deleter
being initialized, which caused a crash in some circumstances.

While doing this, add a test to validate this in the general case for the
sandbox module given that this is not the first time we spot a similar issue
in it.

Follow up to https://github.com/bazelbuild/bazel/issues/7527.

RELNOTES: None.
PiperOrigin-RevId: 268512558