Avoid complex logic in MockSpawnStrategy for better readability.
Push the necessary logic within each instance of the remote and local
strategies in each test to the tests themselves so that they are
self-contained.

This removes all the conditional logic from the generic MockSpawnStrategy,
which was there to prevent code duplication but made reasoning about each
test very difficult.

RELNOTES: None.
PiperOrigin-RevId: 268691053