bazel syntax: simplify comprehensions
This change merges {List,Dict,Abstract}Comprehension into a
single concrete class. The "k: v" body in a dict comprehension
{k: v for vars in iterable} is now represented as
DictionaryLiteral.Entry.

"Clause" was a poor abstraction: clients need to know that
there are only two instances, If and For, and handle them
specifically.

The doEval method is now a single recursive
function that fits easily in a single page.

We delete the following classes:

  AbstractBuilder
  OutputCollector
  ListComprehension
    Builder
    ListOutputCollector
  Clause.Kind
  DictComprehension
    Builder
    DictOutputCollector

PiperOrigin-RevId: 269185401