HttpConnector.connect: request a function instead of fixed set of headers
Fix the signature of HttpConnector.connect. As the connect method
handles redirects, we cannot have a fixed set of headers, once we
support authentication, as the authentication headers depend on
the host and that might change in a redirect.

This change fixes the header generation in the HttpDownloader,
but does not yet re-add the usage of the auth parmeter in the
Starlark-provided download functions.

Related #9327

Change-Id: Id12aa6a9a790fac6133a4da64baff58f02d36ef4
PiperOrigin-RevId: 269300472