Specify 2.7 instead of 2 for python version
This changes the python setup to prefer an executable named `python2.7`
over `python2`. This fixes an issue on macOS where it doesn't ship with
a `python2` executable, but you can install one from homebrew, but
really you'd prefer to be using the system `python2.7` binary. I assume
that this isn't a huge issue because bazel likely doesn't support
versions of python before 2.7.