bazel packages: remove attribute locations
Currently, along with the value of each rule attribute, Blaze records
the optional source location of the identifier of the named
argument that provides the attribute's value, such as 'srcs' in
cc_binary(..., srcs=[...]). This helps blaze emit more precise errors
for problems in a particular attribute, especially for rule
instantiation calls that span many lines.

However, this feature requires a power of introspection not found in
most programming languages: the ability to walk up the call stack to
find the caller and then inspect its syntax tree. As specified,
the Starlark calling convention provides the callee with only the
thread, the values of the positional arguments, and the names and
values of the named arguments; it does not provide source syntax.

The source syntax parameter is an obstacle to upcoming evaluator
optimizations, which will compile syntax trees to a more compact,
efficient, and easily serialized executable representation.

Therefore, this change removes the feature.
All attribute errors now report the location of the rule as a whole.

On the plus side, this change should reduce memory usage of loaded packages.

Follow-up changes will remove the FuncallExpression parameter from
the Starlark function call API once all uses have been eliminated.

PiperOrigin-RevId: 269832751