Try to bring some more clarity into DynamicSpawnStrategyTest.
The tests derived from strategyWaitsForBothSpawnsToFinish were really
difficult to understand due to the booleans passed to this function
and the various state checks based on them. I haven't been able to
eliminate the booleans without a lot of code duplication, but moving
the logic to verify their impact on execution to a per-test lambda
makes things slightly easier to reason about.

As part of the reverse-engineering of what this is supposed to do, this
sprinkles comments throughout the tests based on my understanding so far.

And while here (which is the smaller thing that started this CL),
eliminate the booleans from the strategyPropagatesException helper
function.

These changes fix some pre-existing flakiness in this test related to
bad interrupt handling.

RELNOTES: None.
PiperOrigin-RevId: 269885062