Make the dynamic spawn scheduler use cross-cancellation between spawns.
This modifies the dynamic spawn scheduler so that each pair of concurrent
spawns has the ability to explicitly cancel its corresponding spawn as
desired, without relying on each spawn to abort execution before writing
to the output tree.

This is necessary so that we can later allow the remote branch of the
spawn to interrupt the local branch -- but only once all artifacts have
been downloaded. Such feature will come later.

This change is very intrusive to the point where I consider it a
different implementation of the DynamicSpawnStrategy with possible new
concurrency bugs. As a result, I'm keeping the old implementation around
behind a --legacy_spawn_scheduler flag (which defaults to true for now to
preserve the old behavior). The goal is, of course, to remove the legacy
implementation once we are confident with the new one.

As a side-effect, the changes in how exceptions are handled allow crashes
to propagate properly through the dynamic scheduler. They were previously
being coerced into ExecException, which made debugging more difficult.

RELNOTES: None.
PiperOrigin-RevId: 270039481