bazel skyframe: break dependencies on SkylarkImport
Currently, at parse time we construct a SkylarkImport for each load
statement, applying various Bazel-specific checks to the string,
such as interpreting it as a label that ends in .bzl.
These are not concepts in the Starlark spec, so they do not
belong in lib.syntax.

This change parses and validates load strings in skyframe's package loading recursion.
These checks duplicate the existing ones done in BuildFileAST,
but the latter ones will be removed in a follow-up.
BuildFileAST no longer exposes SkylarkImport in its API,
and the SkylarkImport class is now private, and will
itself be deleted in a follow-up.

SkylarkImportLookupFunction is simplified by parsing load strings as
labels, once, with the correct repo mapping, rather than by fixing up
incorrect labels created too early.

PiperOrigin-RevId: 270344600